+++ packages/eas-cli/src/commands/update/list.ts
@@ -1,4 +1,3 @@
-import { getConfig } from '@expo/config';
 import { Flags } from '@oclif/core';
 import assert from 'assert';
 import chalk from 'chalk';
@@ -8,6 +7,7 @@ import EasCommand from '../../commandUtils/EasCommand';
 import { ViewAllUpdatesQuery } from '../../graphql/generated';
 import { UpdateQuery } from '../../graphql/queries/UpdateQuery';
 import Log from '../../log';
+import { getExpoConfig } from '../../project/expoConfig';
 import { findProjectRootAsync, getProjectIdAsync } from '../../project/projectUtils';
 import { promptAsync } from '../../prompts';
 import { FormatUpdateParameter, UPDATE_COLUMNS, formatUpdate } from '../../update/utils';
@@ -50,7 +50,7 @@ export default class BranchView extends EasCommand {
     }
 
     const projectDir = await findProjectRootAsync();
-    const { exp } = getConfig(projectDir, { skipSDKVersionRequirement: true });
+    const exp = getExpoConfig(projectDir);
     const projectId = await getProjectIdAsync(exp);
 
     let updateGroupDescriptions: UpdateGroupDescription[];