+++ packages/eas-cli/src/submit/submit.ts
@@ -78,7 +78,7 @@ function printInstructionsForIosSubmission(submission: SubmissionFragment): void
       '- It usually takes about 5-10 minutes depending on how busy Apple servers are.',
       // ascAppIdentifier should be always available for ios submissions but check it anyway
       submission.iosConfig?.ascAppIdentifier &&
-        `- When it’s done, you can see your build here: ${link(
+        `- When it's done, you can see your build here: ${link(
           `https://appstoreconnect.apple.com/apps/${submission.iosConfig?.ascAppIdentifier}/appstore/ios`
         )}`,
     ].join('\n');